frontend/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
bcrypt==4.1.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gTTS==2.5.4
httplib2==0.30.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.3
openai-whisper==20250625
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.1.1
psycopg2-binary==2.9.9
pymongo==4.6.1
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
